+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Statistics and About pages added
+- Added annotations search in the home page via FunFam id, Pfam id  (seq and model), Pfam name (seq and model), Pfam descr  (only model).
 - Added filters based on annotations existence.
 - Nav link to the FTP site with seed MSA files and the MGnifams HMM library.
 - Home intro paragraph describing the MGnifams pipeline input data from the latest release of MGnify proteins (2024_04).